Move search to a seperate thread.   Add proxy
