+++ src/main/resources/application.properties
@@ -9,7 +9,11 @@ spring.datasource.password = pass
 spring.jpa.hibernate.ddl-auto=none
 
 # search criteria
-server.search.maxnodes = 500
+server.search.maxnodes = 50
 server.search.maxphotos = 10000
 server.search.baseurl = http://www.stampinup.com/ECWeb/ItemList.aspx?categoryid=102401
 server.version = 0.1
+
+# proxy settings
+#server.proxy.host = http://web-proxy.corp.hpecorp.net
+#server.proxy.port = 8080
